+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * [Hybrid Query] Propagate setMinCompetitiveScore to sub-query scorers in HybridBulkScorer to enable WAND block-level skipping ([#1831](https://github.com/opensearch-project/neural-search/pull/1831))
 
 ### Bug Fixes
+* Support inner_hits for batch semantic highlighting so highlights are produced on nested fields retrieved via `inner_hits` ([#1842](https://github.com/opensearch-project/neural-search/pull/1842))
 * Fix semantic highlighter crash on documents with missing highlighted fields ([#1810](https://github.com/opensearch-project/neural-search/pull/1810))
 * [Text Chunking] Fix text chunking processor ignoring index max_token_count setting when ingesting via alias ([#1803](https://github.com/opensearch-project/neural-search/pull/1803))